EGPS_Web_Scraper_daily/requirements.txt
requests
beautifulsoup4
lxml
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
